requirements.txt
prophet==1.1.6
pandas==2.2.3
flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
psycopg2-binary==2.9.10
